# perf: Push down slice with non-zero offset to Parquet