--- a/crates/polars-io/src/parquet/read/read_impl.rs
+++ b/crates/polars-io/src/parquet/read/read_impl.rs
@@ -23,6 +23,7 @@ use crate::mmap::{MmapBytesReader, ReaderBytes};
 use crate::parquet::metadata::FileMetaDataRef;
 use crate::predicates::{apply_predicate, PhysicalIoExpr};
 use crate::utils::get_reader_bytes;
+use crate::utils::slice::split_slice_at_file;
 use crate::RowIndex;
 
 #[cfg(debug_assertions)]
@@ -156,7 +157,7 @@ fn rg_to_dfs(
     previous_row_count: &mut IdxSize,
     row_group_start: usize,
     row_group_end: usize,
-    remaining_rows: &mut usize,
+    slice: (usize, usize),
     file_metadata: &FileMetaData,
     schema: &ArrowSchemaRef,
     predicate: Option<&dyn PhysicalIoExpr>,
@@ -172,7 +173,7 @@ fn rg_to_dfs(
             previous_row_count,
             row_group_start,
             row_group_end,
-            remaining_rows,
+            slice,
             file_metadata,
             schema,
             predicate,
@@ -188,7 +189,7 @@ fn rg_to_dfs(
             row_group_start,
             row_group_end,
             previous_row_count,
-            remaining_rows,
+            slice,
             file_metadata,
             schema,
             predicate,
@@ -207,7 +208,7 @@ fn rg_to_dfs_optionally_par_over_columns(
     previous_row_count: &mut IdxSize,
     row_group_start: usize,
     row_group_end: usize,
-    remaining_rows: &mut usize,
+    slice: (usize, usize),
     file_metadata: &FileMetaData,
     schema: &ArrowSchemaRef,
     predicate: Option<&dyn PhysicalIoExpr>,
@@ -219,14 +220,21 @@ fn rg_to_dfs_optionally_par_over_columns(
 ) -> PolarsResult<Vec<DataFrame>> {
     let mut dfs = Vec::with_capacity(row_group_end - row_group_start);
 
+    let mut n_rows_processed: usize = (0..row_group_start)
+        .map(|i| file_metadata.row_groups[i].num_rows())
+        .sum();
+    let slice_end = slice.0 + slice.1;
+
     for rg_idx in row_group_start..row_group_end {
         let md = &file_metadata.row_groups[rg_idx];
+        let rg_slice =
+            split_slice_at_file(&mut n_rows_processed, md.num_rows(), slice.0, slice_end);
         let current_row_count = md.num_rows() as IdxSize;
 
         if use_statistics
             && !read_this_row_group(predicate, &file_metadata.row_groups[rg_idx], schema)?
         {
-            *previous_row_count += current_row_count;
+            *previous_row_count += rg_slice.1 as IdxSize;
             continue;
         }
         // test we don't read the parquet file if this env var is set
@@ -235,13 +243,20 @@ fn rg_to_dfs_optionally_par_over_columns(
             assert!(std::env::var("POLARS_PANIC_IF_PARQUET_PARSED").is_err())
         }
 
-        let projection_height = (*remaining_rows).min(md.num_rows());
+        let idx_to_series_projection_height = rg_slice.0 + rg_slice.1;
         let columns = if let ParallelStrategy::Columns = parallel {
             POOL.install(|| {
                 projection
                     .par_iter()
                     .map(|column_i| {
-                        column_idx_to_series(*column_i, md, projection_height, schema, store)
+                        column_idx_to_series(
+                            *column_i,
+                            md,
+                            idx_to_series_projection_height,
+                            schema,
+                            store,
+                        )
+                        .map(|s| s.slice(rg_slice.0 as i64, rg_slice.1))
                     })
                     .collect::<PolarsResult<Vec<_>>>()
             })?
@@ -249,33 +264,34 @@ fn rg_to_dfs_optionally_par_over_columns(
             projection
                 .iter()
                 .map(|column_i| {
-                    column_idx_to_series(*column_i, md, projection_height, schema, store)
+                    column_idx_to_series(
+                        *column_i,
+                        md,
+                        idx_to_series_projection_height,
+                        schema,
+                        store,
+                    )
+                    .map(|s| s.slice(rg_slice.0 as i64, rg_slice.1))
                 })
                 .collect::<PolarsResult<Vec<_>>>()?
         };
-
-        *remaining_rows -= projection_height;
 
         let mut df = unsafe { DataFrame::new_no_checks(columns) };
         if let Some(rc) = &row_index {
             df.with_row_index_mut(&rc.name, Some(*previous_row_count + rc.offset));
         }
 
-        materialize_hive_partitions(
-            &mut df,
-            schema.as_ref(),
-            hive_partition_columns,
-            projection_height,
-        );
+        materialize_hive_partitions(&mut df, schema.as_ref(), hive_partition_columns, rg_slice.1);
         apply_predicate(&mut df, predicate, true)?;
 
         *previous_row_count += current_row_count;
         dfs.push(df);
 
-        if *remaining_rows == 0 {
+        if *previous_row_count as usize >= slice_end {
             break;
         }
     }
+
     Ok(dfs)
 }
 
@@ -286,7 +302,7 @@ fn rg_to_dfs_par_over_rg(
     row_group_start: usize,
     row_group_end: usize,
     previous_row_count: &mut IdxSize,
-    remaining_rows: &mut usize,
+    slice: (usize, usize),
     file_metadata: &FileMetaData,
     schema: &ArrowSchemaRef,
     predicate: Option<&dyn PhysicalIoExpr>,
@@ -296,28 +312,32 @@ fn rg_to_dfs_par_over_rg(
     hive_partition_columns: Option<&[Series]>,
 ) -> PolarsResult<Vec<DataFrame>> {
     // compute the limits per row group and the row count offsets
-    let row_groups = file_metadata
-        .row_groups
-        .iter()
-        .enumerate()
-        .skip(row_group_start)
-        .take(row_group_end - row_group_start)
-        .map(|(rg_idx, rg_md)| {
-            let row_count_start = *previous_row_count;
-            let num_rows = rg_md.num_rows();
-            *previous_row_count += num_rows as IdxSize;
-            let projection_height = (*remaining_rows).min(num_rows);
-            *remaining_rows -= projection_height;
+    let mut row_groups = Vec::with_capacity(row_group_end - row_group_start);
 
-            (rg_idx, rg_md, projection_height, row_count_start)
-        })
-        .collect::<Vec<_>>();
+    let mut n_rows_processed: usize = (0..row_group_start)
+        .map(|i| file_metadata.row_groups[i].num_rows())
+        .sum();
+    let slice_end = slice.0 + slice.1;
+
+    for i in row_group_start..row_group_end {
+        let row_count_start = *previous_row_count;
+        let rg_md = &file_metadata.row_groups[i];
+        let rg_slice =
+            split_slice_at_file(&mut n_rows_processed, rg_md.num_rows(), slice.0, slice_end);
+        *previous_row_count += rg_slice.1 as IdxSize;
+
+        if rg_slice.1 == 0 {
+            continue;
+        }
+
+        row_groups.push((i, rg_md, rg_slice, row_count_start));
+    }
 
     let dfs = POOL.install(|| {
         row_groups
             .into_par_iter()
-            .map(|(rg_idx, md, projection_height, row_count_start)| {
-                if projection_height == 0
+            .map(|(rg_idx, md, slice, row_count_start)| {
+                if slice.1 == 0
                     || use_statistics
                         && !read_this_row_group(
                             predicate,
@@ -336,7 +356,8 @@ fn rg_to_dfs_par_over_rg(
                 let columns = projection
                     .iter()
                     .map(|column_i| {
-                        column_idx_to_series(*column_i, md, projection_height, schema, store)
+                        column_idx_to_series(*column_i, md, slice.0 + slice.1, schema, store)
+                            .map(|x| x.slice(slice.0 as i64, slice.1))
                     })
                     .collect::<PolarsResult<Vec<_>>>()?;
 
@@ -350,7 +371,7 @@ fn rg_to_dfs_par_over_rg(
                     &mut df,
                     schema.as_ref(),
                     hive_partition_columns,
-                    projection_height,
+                    slice.1,
                 );
                 apply_predicate(&mut df, predicate, false)?;
 
@@ -364,7 +385,7 @@ fn rg_to_dfs_par_over_rg(
 #[allow(clippy::too_many_arguments)]
 pub fn read_parquet<R: MmapBytesReader>(
     mut reader: R,
-    mut limit: usize,
+    slice: (usize, usize),
     projection: Option<&[usize]>,
     reader_schema: &ArrowSchemaRef,
     metadata: Option<FileMetaDataRef>,
@@ -375,7 +396,7 @@ pub fn read_parquet<R: MmapBytesReader>(
     hive_partition_columns: Option<&[Series]>,
 ) -> PolarsResult<DataFrame> {
     // Fast path.
-    if limit == 0 {
+    if slice.1 == 0 {
         return Ok(materialize_empty_df(
             projection,
             reader_schema,
@@ -430,7 +451,7 @@ pub fn read_parquet<R: MmapBytesReader>(
         &mut 0,
         0,
         n_row_groups,
-        &mut limit,
+        slice,
         &file_metadata,
         reader_schema,
         predicate,
@@ -511,28 +532,51 @@ impl RowGroupFetcher {
 pub(super) fn compute_row_group_range(
     row_group_start: usize,
     row_group_end: usize,
-    limit: usize,
+    slice: (usize, usize),
     row_groups: &[RowGroupMetaData],
-) -> usize {
-    let mut row_group_end_truncated = row_group_start;
-    let mut acc_row_count = 0;
+) -> std::ops::Range<usize> {
+    let mut start = row_group_start;
+    let mut cum_rows: usize = (0..row_group_start).map(|i| row_groups[i].num_rows()).sum();
+    let row_group_end = row_groups.len().min(row_group_end);
 
-    #[allow(clippy::needless_range_loop)]
-    for rg_i in row_group_start..(std::cmp::min(row_group_end, row_groups.len())) {
-        if acc_row_count >= limit {
+    loop {
+        if start == row_group_end {
             break;
         }
-        row_group_end_truncated = rg_i + 1;
-        acc_row_count += row_groups[rg_i].num_rows();
+
+        cum_rows += row_groups[start].num_rows();
+
+        if cum_rows >= slice.0 {
+            break;
+        }
+
+        start += 1;
     }
-    row_group_end_truncated
+
+    let slice_end = slice.0 + slice.1;
+    let mut end = (1 + start).min(row_group_end);
+
+    loop {
+        if end == row_group_end {
+            break;
+        }
+
+        if cum_rows >= slice_end {
+            break;
+        }
+
+        cum_rows += row_groups[end].num_rows();
+        end += 1;
+    }
+
+    start..end
 }
 
 pub struct BatchedParquetReader {
     // use to keep ownership
     #[allow(dead_code)]
     row_group_fetcher: RowGroupFetcher,
-    limit: usize,
+    slice: (usize, usize),
     projection: Arc<[usize]>,
     schema: ArrowSchemaRef,
     metadata: FileMetaDataRef,
@@ -557,7 +601,7 @@ impl BatchedParquetReader {
         row_group_fetcher: RowGroupFetcher,
         metadata: FileMetaDataRef,
         schema: ArrowSchemaRef,
-        limit: usize,
+        slice: (usize, usize),
         projection: Option<Vec<usize>>,
         predicate: Option<Arc<dyn PhysicalIoExpr>>,
         row_index: Option<RowIndex>,
@@ -589,7 +633,7 @@ impl BatchedParquetReader {
 
         Ok(BatchedParquetReader {
             row_group_fetcher,
-            limit,
+            slice,
             projection,
             schema,
             metadata,
@@ -609,10 +653,6 @@ impl BatchedParquetReader {
         })
     }
 
-    pub fn limit_reached(&self) -> bool {
-        self.limit == 0
-    }
-
     pub fn schema(&self) -> &ArrowSchemaRef {
         &self.schema
     }
@@ -626,7 +666,7 @@ impl BatchedParquetReader {
     }
 
     pub async fn next_batches(&mut self, n: usize) -> PolarsResult<Option<Vec<DataFrame>>> {
-        if self.limit == 0 && self.has_returned {
+        if self.rows_read as usize == self.slice.0 + self.slice.1 && self.has_returned {
             return if self.chunks_fifo.is_empty() {
                 Ok(None)
             } else {
@@ -638,28 +678,30 @@ impl BatchedParquetReader {
 
         let mut skipped_all_rgs = false;
         // fill up fifo stack
-        if self.row_group_offset < self.n_row_groups && self.chunks_fifo.len() < n {
+        if (self.rows_read as usize) < self.slice.0 + self.slice.1
+            && self.row_group_offset < self.n_row_groups
+            && self.chunks_fifo.len() < n
+        {
             // Ensure we apply the limit on the metadata, before we download the row-groups.
-            let row_group_start = self.row_group_offset;
-            let row_group_end = compute_row_group_range(
-                row_group_start,
-                row_group_start + n,
-                self.limit,
+            let row_group_range = compute_row_group_range(
+                self.row_group_offset,
+                self.row_group_offset + n,
+                self.slice,
                 &self.metadata.row_groups,
             );
 
             let store = self
                 .row_group_fetcher
-                .fetch_row_groups(row_group_start..row_group_end)
+                .fetch_row_groups(row_group_range.clone())
                 .await?;
 
             let mut dfs = match store {
                 ColumnStore::Local(_) => rg_to_dfs(
                     &store,
                     &mut self.rows_read,
-                    row_group_start,
-                    row_group_end,
-                    &mut self.limit,
+                    row_group_range.start,
+                    row_group_range.end,
+                    self.slice,
                     &self.metadata,
                     &self.schema,
                     self.predicate.as_deref(),
@@ -680,7 +722,6 @@ impl BatchedParquetReader {
 
                     // Make everything 'static.
                     let mut rows_read = self.rows_read;
-                    let mut limit = self.limit;
                     let row_index = self.row_index.clone();
                     let predicate = self.predicate.clone();
                     let schema = self.schema.clone();
@@ -689,14 +730,15 @@ impl BatchedParquetReader {
                     let projection = self.projection.clone();
                     let use_statistics = self.use_statistics;
                     let hive_partition_columns = self.hive_partition_columns.clone();
+                    let slice = self.slice;
 
                     let f = move || {
                         let dfs = rg_to_dfs(
                             &store,
                             &mut rows_read,
-                            row_group_start,
-                            row_group_end,
-                            &mut limit,
+                            row_group_range.start,
+                            row_group_range.end,
+                            slice,
                             &metadata,
                             &schema,
                             predicate.as_deref(),
@@ -708,7 +750,7 @@ impl BatchedParquetReader {
                         );
 
                         // Don't unwrap send attempt - async task could be cancelled.
-                        let _ = tx.send((dfs, rows_read, limit));
+                        let _ = tx.send((dfs, rows_read));
                     };
 
                     // Spawn the task and wait on it asynchronously.
@@ -721,10 +763,8 @@ impl BatchedParquetReader {
                         POOL.spawn(f);
                     };
 
-                    let (dfs, rows_read, limit) = rx.await.unwrap();
-
+                    let (dfs, rows_read) = rx.await.unwrap();
                     self.rows_read = rows_read;
-                    self.limit = limit;
                     dfs
                 },
             }?;