--- a/crates/polars-io/src/utils/mod.rs
+++ b/crates/polars-io/src/utils/mod.rs
@@ -3,6 +3,7 @@ mod other;
 
 pub use compression::is_compressed;
 pub use other::*;
+pub mod slice;
 
 pub const URL_ENCODE_CHAR_SET: &percent_encoding::AsciiSet = &percent_encoding::CONTROLS
     .add(b'/')