--- a/crates/polars-mem-engine/src/executors/scan/mod.rs
+++ b/crates/polars-mem-engine/src/executors/scan/mod.rs
@@ -110,7 +110,10 @@ pub(crate) struct AnonymousScanExec {
 impl Executor for AnonymousScanExec {
     fn execute(&mut self, state: &mut ExecutionState) -> PolarsResult<DataFrame> {
         let mut args = AnonymousScanArgs {
-            n_rows: self.file_options.n_rows,
+            n_rows: self.file_options.slice.map(|x| {
+                assert_eq!(x.0, 0);
+                x.1
+            }),
             with_columns: self.file_options.with_columns.clone(),
             schema: self.file_info.schema.clone(),
             output_schema: self.output_schema.clone(),