… crates/polars-lazy/src/tests/io.rs
… crates/polars-lazy/src/tests/optimization_checks.rs
… crates/polars-mem-engine/src/executors/scan/parquet.rs
… crates/polars-pipe/src/executors/sources/csv.rs
… crates/polars-pipe/src/executors/sources/parquet.rs
… crates/polars-plan/src/plans/builder_dsl.rs
… crates/polars-plan/src/plans/ir/format.rs
… crates/polars-plan/src/plans/optimizer/predicate_pushdown/mod.rs
… crates/polars-plan/src/plans/optimizer/slice_pushdown_lp.rs
… crates/polars-plan/src/plans/options.rs
… py-polars/src/dataframe/io.rs
… py-polars/src/lazyframe/visitor/nodes.rs
… py-polars/tests/unit/io/test_hive.py
… py-polars/tests/unit/io/test_lazy_parquet.py