Closes https://github.com/pola-rs/polars/issues/17905

Scanning the example parquet file from https://github.com/pola-rs/polars/issues/17905#issue-2433407593:
```python
pq_file = "demo.parquet"
lf = pl.scan_parquet(pq_file)
shape = (lf.select(pl.len()).collect().item(), lf.collect_schema().len())
print(shape)
print(lf.tail(1).collect())

```

Release 1.3.0:
```python
python .env/y.py  9.10s user 4.16s system 500% cpu 2.649 total
```
This branch (on a debug build):
```python
python .env/y.py  0.66s user 0.04s system 95% cpu 0.728 total
```