@coastalwhite 2024-08-01
I think this should probably also make use of the `Filter::Range`, but I can do that afterwards.