@wence- 2024-08-02
@nameexhaustion Sorry, I didn't catch this in the python IR, _technically_ this should have bumped the `NodeTraverser` major version since the `file_options` object changed. However, since we already had a bump between 1.3 and 1.4 no harm done. We only recently added this feature so it's not obvious that one had to do it.